+++ src/async/Workflows/index.js
@@ -1,0 +1,4 @@
+module.exports.Version = require("./Version");
+module.exports.AbstractWorkflow = require("./AbstractWorkflow");
+module.exports.Workflow = require("./Workflow");
+module.exports.workflowManager = require("./WorkflowManager");